Cleaned tft_ace code; 'Clear' tags made USER_DEFINED
