+++ tft_ace/config.cpp
@@ -12,18 +12,7 @@
 };
 
 class CfgFunctions {
-    class tft_ace {
-        class tft_ace {
-            file = "tft_ace\functions";
-            class scream;
-            class getIn;
-            class canTag;
-            class tag;
-            class tagClear;
-            class markClear;
-            class markerSize;
-        };
-    };
+    #include "functions\CfgFunctions.hpp"
 };
 
 class CfgVehicles {
@@ -34,7 +23,7 @@ class CfgVehicles {
                 class ACE_tagWallArtistic {
                     displayName = "Go artistic";
                     condition = "[_player] call tft_ace_fnc_canTag";
-                    statement = "[_player, 5] call tft_ace_fnc_tag";
+                    statement = "[_player, 'artistic', 5] call tft_ace_fnc_tag";
                     showDisabled = 0;
                     priority = 3;
                     icon = "\tft_ace\UI\icons\retag.paa";
@@ -42,10 +31,10 @@ class CfgVehicles {
                 class ACE_tagClear {
                     displayName = "Mark Clear";
                     condition = "[_player] call ACE_tagging_fnc_checkTaggable && {'ACE_SpraypaintGreen' in items _player}";
-                    statement = "[_player,3] call tft_ace_fnc_tagClear; _player call tft_ace_fnc_markClear;";
+                    statement = "[_player, 'clear', 3] call tft_ace_fnc_tag; [_player] call tft_ace_fnc_markClear;";
                     showDisabled = 0;
                     priority = 3;
-                    icon = "\tft_ace\UI\icons\clearIcon_ca.paa";
+                    icon = "\tft_ace\UI\icons\clearIcon.paa";
                 };
             };
         };
@@ -73,12 +62,13 @@ class CfgMarkerClasses {
 class CfgMarkers {
     class tft_clearMarker {
         name = "Building Clear";
-        icon = "\tft_ace\UI\icons\clearIcon_ca.paa";
+        icon = "\tft_ace\UI\icons\clearIcon.paa";
         markerClass = "tft_markers";
-        color[] = {1, 1, 1, 1};
+        color[] = {"(profilenamespace getvariable ['Map_BLUFOR_R',0])","(profilenamespace getvariable ['Map_BLUFOR_G',1])","(profilenamespace getvariable ['Map_BLUFOR_B',1])","(profilenamespace getvariable ['Map_BLUFOR_A',0.8])"};
         size = 20;
         scope = 2;
         scopeCurator = 2;
+        texture = "\tft_ace\UI\icons\clearIcon.paa";
         shadow = 0;
     };
 };
