+++ tft_ace/functions/CfgFunctions.hpp
@@ -1,0 +1,10 @@
+class tft_ace {
+    class tft_ace {
+        file = "tft_ace\functions";
+        class scream;
+        class getIn;
+        class canTag;
+        class tag;
+        class markClear;
+    };
+};
