+++ tft_ace/functions/fn_canTag.sqf
@@ -1,22 +1,20 @@
 /*
 Author: YoursTruly
-[Description]
+Check if it's possible to create a tag here.
 
 * Arguments:
-* 0: player
+* 0: Unit trying to tag <OBJECT>
 *
 * Return Value:
-* The return value <BOOL>
+* -
 
 * Example:
-* ["something", player] call ace_common_fnc_imanexample
-
-* Public: [Yes/No]
+* [player] call tft_ace_fnc_canTag
 */
 
 params ["_player"];
 
-_hasSpray = count (['ACE_SpraypaintBlack','ACE_SpraypaintRed','ACE_SpraypaintGreen','ACE_SpraypaintBlue'] arrayIntersect items _player) > 0;
-_closeToTag = [_player] call ace_tagging_fnc_checkTaggable;
+private _hasSpray = count (['ACE_SpraypaintBlack','ACE_SpraypaintRed','ACE_SpraypaintGreen','ACE_SpraypaintBlue'] arrayIntersect items _player) > 0;
+private _taggableArea = [_player] call ace_tagging_fnc_checkTaggable;
 
-_hasSpray && {_closeToTag}
+_hasSpray && {_taggableArea}
