+++ tft_ace/functions/fn_scream.sqf
@@ -1,28 +1,22 @@
 /*
- * Author: [Name of Author(s)]
- * [Description]
+ * Author: YoursTruly
+ * Makes a player scream when he goes unconscious.
  *
  * Arguments:
- * 0: The first argument <STRING>
- * 1: The second argument <OBJECT>
+ * 0: Source of a scream <OBJECT>
  *
  * Return Value:
- * The return value <BOOL>
+ * -
  *
  * Example:
- * ["something", player] call ace_common_fnc_imanexample
- *
- * Public: [Yes/No]
+ * player call tft_ace_fnc_scream
  */
-
-private ["_unit", "_screams", "_scream", "_position"];
-
-_unit = _this;
+private _unit = _this;
 
 if(!local _unit) exitWith {};
 if!(_unit in playableUnits) exitWith {};
 
-_screams = [
+private _screams = [
   "A3\sounds_f\characters\human-sfx\Person0\P0_hit_01.wss",
   "A3\sounds_f\characters\human-sfx\Person0\P0_hit_02.wss",
   "A3\sounds_f\characters\human-sfx\Person0\P0_hit_03.wss",
@@ -71,10 +65,10 @@ _screams = [
   "A3\sounds_f\characters\human-sfx\Person3\P3_hit_13.wss"
 ];
 
-_position = _unit modelToWorld (_unit selectionPosition "Head");
+private _position = _unit modelToWorld (_unit selectionPosition "Head");
 _position set [2, (_position select 2) + ((getPosASLW _unit select 2) - (getPosATL _unit select 2) max 0)];
 
-_scream = _screams select (floor random (count _screams));
+private _scream = _screams select (floor random (count _screams));
 
 playSound3D [
   _scream,
