+++ tft_ace/functions/fn_tag.sqf
@@ -1,21 +1,20 @@
 /*
-* Author: [Name of Author(s)]
-* [Description]
+* Author: YoursTruly
+* Create a tag choosing one of the random tags.
 *
 * Arguments:
-* 0: The first argument <STRING>
-* 1: The second argument <OBJECT>
+* 0: Unit placing a tag <OBJECT>
+* 1: Prefix for a tag name <STRING>
+* 2: Number of tags with given prefixes to choose from <NUMBER>
 *
 * Return Value:
-* The return value <BOOL>
+* -
 *
 * Example:
-* ["something", player] call ace_common_fnc_imanexample
-*
-* Public: [Yes/No]
+* [player, "tag_", 4] call tft_ace_fnc_tag
 */
 
-params ["_player", "_tagNum"];
+params ["_player", "_tagPrefix", "_tagNum"];
 
-_tag = 'tft_ace\UI\tags\' + str (floor (random _tagNum)) + '.paa';
+private _tag = "tft_ace\UI\tags\" + _tagPrefix + str (floor (random _tagNum)) + ".paa";
 [_player, _tag] call ace_tagging_fnc_tag;
